requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.25.2
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
pytest==8.0.0
pytest-asyncio==0.23.5
aiosqlite==0.19.0
